Fix data type for mouse wheel events

+++ source/gloperate-qt/source/viewer/RenderWindow.cpp
@@ -144,8 +144,8 @@ void RenderWindow::mouseReleaseEvent(QMouseEvent * event)
 void RenderWindow::wheelEvent(QWheelEvent * event)
 {
     m_surface->onMouseWheel(
-        glm::ivec2( event->orientation() == Qt::Vertical ? 0 : event->delta(),
-                    event->orientation() == Qt::Vertical ? event->delta() : 0 ),
+        glm::vec2( event->orientation() == Qt::Vertical ? 0.0f : (float)event->delta(),
+                   event->orientation() == Qt::Vertical ? (float)event->delta() : 0.0f ),
         glm::ivec2( (int)(event->x() * devicePixelRatio()),
                     (int)(event->y() * devicePixelRatio()) )
     );
